refactor: clarify HID accumulator ownership
